requirements.txt
torch
torchvision
zennit-crp
pytorch-lightning
pandas
opencv-python
wandb
scikit-learn
umap-learn
timm
safetensors
